requirements.txt
chromadb>=0.4.22
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
boto3>=1.28.0
